+++ tests/EventDispatcherTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Konecnyjakub\EventDispatcher;
 
+use Konecnyjakub\EventDispatcher\Events\AbstractEvent;
 use Konecnyjakub\EventDispatcher\Events\Event;
 use Konecnyjakub\EventDispatcher\Events\TestStoppableEvent;
 use MyTester\Attributes\TestSuite;
@@ -22,6 +23,19 @@ final class EventDispatcherTest extends TestCase
         $eventDispatcher = new EventDispatcher($listenerProvider);
         $this->assertSame($event, $eventDispatcher->dispatch($event));
         $this->assertSame(1, $var);
+
+        $event = new Event();
+        $var = 0;
+        $listenerProvider = new AutoListenerProvider();
+        $listenerProvider->addListener(function (Event $event) use (&$var): void {
+            $var++;
+        });
+        $listenerProvider->addListener(function (AbstractEvent $event) use (&$var): void {
+            $var += 2;
+        });
+        $eventDispatcher = new EventDispatcher($listenerProvider);
+        $this->assertSame($event, $eventDispatcher->dispatch($event));
+        $this->assertSame(3, $var);
 
         $event = new TestStoppableEvent();
         $var = 0;