+++ tests/Events/AbstractEvent.php
@@ -3,6 +3,6 @@ declare(strict_types=1);
 
 namespace Konecnyjakub\EventDispatcher\Events;
 
-final class Event extends AbstractEvent
+abstract class AbstractEvent
 {
 }